iot_device/requirements.txt
paho-mqtt
requests
pytz
python-dotenv